+++ src/Screens/MusicSelect/Panels/Panel.cpp
@@ -97,9 +97,9 @@ namespace MusicSelect {
         selected_chart.reset();
     }
 
-    std::optional<SongDifficulty> SongPanel::get_selected_difficulty() const {
+    std::optional<Data::SongDifficulty> SongPanel::get_selected_difficulty() const {
         if (selected_chart) {
-            return SongDifficulty{*m_song, *selected_chart};
+            return Data::SongDifficulty{*m_song, *selected_chart};
         } else {
             return {};
         }